% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calc_streak.R
\name{calc_streak}
\alias{calc_streak}
\title{calc_streak}
\usage{
calc_streak(x, streakvar = "H")
}
\arguments{
\item{x}{A sequence of two values (default is  \code{"H"} and \code{"M"}) indicating a sequence
  of possible outcomes for a binary variable. For example,  \code{"H"} and \code{"M"} could indicate
  hits and misses in basketball shots, or \code{"heads"} and \code{"tails"} could indicate coin flips.}

\item{streakvar}{The value that you are interested in studying. Defaults to \code{"H"} but could be
 changed to \code{"heads"} or whatever other response you were interested in studying.}
}
\value{
A vector of the streak lengths. A streak is defined as a series of
  of repeated instances of the \code{streakvar}, followed by the other binary outcome.
  For example, a series of hits (\code{"H"}) followed by a miss (\code{"M"}).
}
\description{
calc_streak
}
\examples{
sequence <- c("H", "H", "M", "H", "M", "M", "H", "H", "H")
calc_streak(sequence)
}

