% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{cdc}
\alias{cdc}
\title{Behavioral survey}
\format{A data frame with 20,000 rows and 9 variables:
\describe{
   \item{genhlth}{A categorical vector indicating general health, with
   categories excellent, very good, good, fair, and poor}
   \item{exerany}{A categorical vector, 1 if the respondent exercised in
   the past month and 0 otherwise}
   \item{hlthplan}{A categorical vector, 1 if the respondent has some form
   of health coverage and 0 otherwise}
   \item{smoke100}{A categorical vector, 1 if the respondent has smoked at
   least 100 cigarettes in their entire life and 0 otherwise}
   \item{height}{A numerical vector, respondent's height in inches}
   \item{weight}{A numerical vector, respondent's weight in pounds}
   \item{wtdesire}{A numerical vector, respondent's desired weight in pounds}
   \item{age}{A numerical vector, respondent's age in years}
   \item{gender}{A numerical vector, respondent's gender}
}}
\source{
These data are excerpted from the Behavioral Risk Factor Surveillance
System survey conducted by the CDC. For more information, see the
\href{http://www.cdc.gov/brfss/}{BRFSS Web site}.
}
\usage{
cdc
}
\description{
This data set is a random sample of 20,000 people from the Behavioral Risk
Factor Surveillance System (BRFSS) survey conducted in 2000. While there
are over 200 questions or variables in the original data set, this one
only includes 9 variables.
}
\details{
The BRFSS is an annual telephone survey of 350,000 people in the United
States collected by the Centers for Disease Control and Prevention (CDC).
As its name implies, the BRFSS is designed to identify risk factors in
the adult population and report emerging health trends. For example,
respondents are asked about their diet and weekly physical activity, their
HIV/AIDS status, possible tobacco use, and even their level of healthcare
coverage. The \href{http://www.cdc.gov/brfss/}{BRFSS Web site} contains a
complete description of the survey, the questions that were asked and even
research results that have been derived from the data.
}
\keyword{datasets}

