% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_single_mean_theo.R
\name{ci_single_mean_theo}
\alias{ci_single_mean_theo}
\title{Helper for the inference() function, not exported}
\usage{
ci_single_mean_theo(y, conf_level, y_name, show_var_types, show_summ_stats,
  show_res, show_eda_plot, show_inf_plot)
}
\description{
Helper for the inference() function, not exported
}
