% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_single_median_sim.R
\name{ci_single_median_sim}
\alias{ci_single_median_sim}
\title{Confidence interval for one median, using bootstrapping}
\usage{
ci_single_median_sim(y, conf_level, y_name, boot_method, nsim, seed,
  show_var_types, show_summ_stats, show_res, show_eda_plot, show_inf_plot)
}
\arguments{
\item{y}{Response variable, can be numerical or categorical}

\item{conf_level}{confidence level, value between 0 and 1}

\item{y_name}{Name of response variable as a character string (passed 
from inference function)}

\item{boot_method}{bootstrap method; "perc" (percentile) or 
"se" (standard error)}

\item{nsim}{number of simulations}

\item{seed}{seed to be set, default is NULL}

\item{show_var_types}{print variable types, set to verbose by default}

\item{show_summ_stats}{print summary stats, set to verbose by default}

\item{show_res}{print results, set to verbose by default}

\item{show_eda_plot}{print EDA plot, set to verbose by default}

\item{show_inf_plot}{print inference plot, set to verbose by default}
}
\description{
Helper for the [inference] function
}
