% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_single_prop_sim.R
\name{ci_single_prop_sim}
\alias{ci_single_prop_sim}
\title{Helper for the inference() function, not exported}
\usage{
ci_single_prop_sim(y, success, conf_level, y_name, boot_method, nsim, seed,
  show_var_types, show_summ_stats, show_res, show_eda_plot, show_inf_plot)
}
\description{
Helper for the inference() function, not exported
}
