% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_two_median_sim.R
\name{ci_two_median_sim}
\alias{ci_two_median_sim}
\title{Helper for the inference() function, not exported}
\usage{
ci_two_median_sim(y, x, conf_level, y_name, x_name, boot_method, nsim, seed,
  show_var_types, show_summ_stats, show_res, show_eda_plot, show_inf_plot)
}
\description{
Helper for the inference() function, not exported
}
