% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_two_prop_sim.R
\name{ci_two_prop_sim}
\alias{ci_two_prop_sim}
\title{Helper for the inference() function, not exported}
\usage{
ci_two_prop_sim(y, x, success, conf_level, x_name, y_name, boot_method, nsim,
  seed, show_var_types, show_summ_stats, show_res, show_eda_plot, show_inf_plot)
}
\description{
Helper for the inference() function, not exported
}
