% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht_many_mean_theo.R
\name{ht_many_mean_theo}
\alias{ht_many_mean_theo}
\title{Hypothesis testing for many means, using ANOVA}
\usage{
ht_many_mean_theo(y, x, null, alternative, sig_level, y_name, x_name,
  show_var_types, show_summ_stats, show_res, show_eda_plot, show_inf_plot)
}
\arguments{
\item{y}{Response variable, can be numerical or categorical}

\item{x}{Explanatory variable, categorical (optional)}

\item{null}{null value for a hypothesis test}

\item{alternative}{direction of the alternative hypothesis; "less","greater", or "twosided"}

\item{sig_level}{significance level, value between 0 and 1 
(used only for ANOVA to determine if posttests are necessary)}

\item{y_name}{Name of response variable as a character string (passed 
from inference function)}

\item{x_name}{Name of explanatory variable as a character string (passed 
from inference function)}

\item{show_var_types}{print variable types, set to verbose by default}

\item{show_summ_stats}{print summary stats, set to verbose by default}

\item{show_res}{print results, set to verbose by default}

\item{show_eda_plot}{print EDA plot, set to verbose by default}

\item{show_inf_plot}{print inference plot, set to verbose by default}
}
\description{
Helper for the `inference()` function
}
