% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht_many_prop_theo.R
\name{ht_many_prop_theo}
\alias{ht_many_prop_theo}
\title{Helper for the inference() function, not exported}
\usage{
ht_many_prop_theo(y, x, x_name, y_name, show_var_types, show_summ_stats,
  show_res, show_eda_plot, show_inf_plot)
}
\description{
Helper for the inference() function, not exported
}
