% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht_single_mean_sim.R
\name{ht_single_mean_sim}
\alias{ht_single_mean_sim}
\title{Helper for the inference() function, not exported}
\usage{
ht_single_mean_sim(y, null, alternative, y_name, nsim, seed, show_var_types,
  show_summ_stats, show_res, show_eda_plot, show_inf_plot)
}
\description{
Helper for the inference() function, not exported
}
