% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht_single_prop_sim.R
\name{ht_single_prop_sim}
\alias{ht_single_prop_sim}
\title{Helper for the inference() function, not exported}
\usage{
ht_single_prop_sim(y, success, null, alternative, nsim, seed, y_name,
  show_var_types, show_summ_stats, show_eda_plot, show_inf_plot, show_res)
}
\description{
Helper for the inference() function, not exported
}
