% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht_single_prop_theo.R
\name{ht_single_prop_theo}
\alias{ht_single_prop_theo}
\title{Helper for the inference() function, not exported}
\usage{
ht_single_prop_theo(y, success, null, alternative, y_name, show_var_types,
  show_summ_stats, show_res, show_eda_plot, show_inf_plot)
}
\description{
Helper for the inference() function, not exported
}
