% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht_two_mean_theo.R
\name{ht_two_mean_theo}
\alias{ht_two_mean_theo}
\title{Helper for the inference() function, not exported}
\usage{
ht_two_mean_theo(y, x, null, alternative, y_name, x_name, show_var_types,
  show_summ_stats, show_res, show_eda_plot, show_inf_plot)
}
\description{
Helper for the inference() function, not exported
}
