% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht_two_prop_theo.R
\name{ht_two_prop_theo}
\alias{ht_two_prop_theo}
\title{Hypothesis testing for two proportions, using CLT-based Z-interval}
\usage{
ht_two_prop_theo(y, x, success, null, alternative, y_name, x_name,
  show_var_types, show_summ_stats, show_res, show_eda_plot, show_inf_plot)
}
\arguments{
\item{y}{Response variable, can be numerical or categorical}

\item{x}{Explanatory variable, categorical (optional)}

\item{success}{which level of the categorical variable to call "success", 
i.e. do inference on}

\item{null}{null value for a hypothesis test}

\item{alternative}{direction of the alternative hypothesis; "less","greater", 
or "twosided"}

\item{y_name}{Name of response variable as a character string (passed 
from inference function)}

\item{x_name}{Name of explanatory variable as a character string (passed 
from inference function)}

\item{show_var_types}{print variable types, set to verbose by default}

\item{show_summ_stats}{print summary stats, set to verbose by default}

\item{show_res}{print results, set to verbose by default}

\item{show_eda_plot}{print EDA plot, set to verbose by default}

\item{show_inf_plot}{print inference plot, set to verbose by default}
}
\description{
Helper for the `inference()` function
}
