% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inference.R
\name{inference}
\alias{inference}
\title{inference}
\usage{
inference(y, x = NULL, est = c("mean", "median", "proportion"),
  success = NULL, order = NULL, conflevel = 0.95, siglevel = 0.05,
  null = NULL, alternative = c("less", "greater", "twosided"),
  type = c("ci", "ht"), method = c("theoretical", "simulation"),
  simdist = FALSE, nsim = 10000, seed = NULL, sum_stats = TRUE,
  eda_plot = TRUE, inf_plot = TRUE, inf_lines = TRUE)
}
\arguments{
\item{y}{}

\item{x}{}

\item{est}{}

\item{success}{}

\item{order}{}

\item{conflevel}{}

\item{siglevel}{}

\item{null}{}

\item{alternative}{}

\item{type}{}

\item{method}{}

\item{simdist}{}

\item{nsim}{}

\item{seed}{}

\item{sum_stats}{}

\item{eda_plot}{}

\item{inf_plot}{}

\item{inf_lines}{}
}
\description{
what does this do?
}

