% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multiLines.R
\name{multiLines}
\alias{multiLines}
\title{multiLines}
\usage{
multiLines(model, ...)
}
\arguments{
\item{model}{an \code{\link{lm}} object}

\item{...}{arguments passed to \code{\link[graphics]{abline}}}
}
\description{
what does this do?
}

