% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oilabs.R
\docType{package}
\name{oilabs}
\alias{oilabs}
\alias{oilabs-package}
\title{oilabs: An R package accompanying the OpenIntro labs.}
\description{
The oilabs package contains the data sets, custom functions, documentation, and templates for the OpenIntro Labs.
}
\details{
In particular, it contains the datasets arbuthnot, atheism, bdims, cdc, evals, kobe, mlb11 and present.

It also contains the functions inference, multiLines, plot_ss and qqnormsim.
}

