% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_ss.R
\name{plot_ss}
\alias{plot_ss}
\title{plot_ss}
\usage{
plot_ss(x, y, showSquares = FALSE, leastSquares = FALSE)
}
\arguments{
\item{x}{}

\item{y}{}

\item{showSquares}{}

\item{leastSquares}{}
}
\description{
what does this do?
}

