% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{present}
\alias{present}
\title{Male and female births in the US}
\format{A data frame with 63 rows and 3 variables:
\describe{
  \item{year}{year, ranging from 1940 to 2002}
  \item{boys}{number of male births}
  \item{girls}{number of female births}
}}
\source{
These data appear in Mathews TJ, and Hamilton BE. 2005. Trend
analysis of the sex ratio at birth in the United States. National Vital
Statistics Reports 53(20):1-17.
}
\usage{
present
}
\description{
Counts of the total number of male and femals births in the United States from
1940 to 2002.
}
\keyword{datasets}

