% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calc_streak.R
\name{calc_streak}
\alias{calc_streak}
\title{calc_streak}
\usage{
calc_streak(x)
}
\arguments{
\item{x}{A sequence of \code{"H"} and \code{"M"} values indicating a sequence
  of hit or missed shots.}
}
\value{
A vector of the streak lengths. A streak is defined as a series of
  hits \code{"H"} followed by a miss \code{"M"}.
}
\description{
calc_streak
}
\examples{
sequence <- c("H", "H", "M", "H", "M", "M", "H", "H", "H")
calc_streak(sequence)
}

