% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht_single_prop_sim.R
\name{ht_single_prop_sim}
\alias{ht_single_prop_sim}
\title{Hypothesis testing for one proportion, using randomization}
\usage{
ht_single_prop_sim(y, success, null, alternative, nsim, seed, y_name,
  show_var_types, show_summ_stats, show_eda_plot, show_inf_plot, show_res)
}
\arguments{
\item{y}{Response variable, can be numerical or categorical}

\item{success}{which level of the categorical variable to call "success",
i.e. do inference on}

\item{null}{null value for a hypothesis test}

\item{alternative}{direction of the alternative hypothesis; "less",
"greater", or "twosided"}

\item{nsim}{number of simulations}

\item{seed}{seed to be set, default is NULL}

\item{y_name}{Name of response variable as a character string (passed
from inference function)}

\item{show_var_types}{print variable types, set to verbose by default}

\item{show_summ_stats}{print summary stats, set to verbose by default}

\item{show_eda_plot}{print EDA plot, set to verbose by default}

\item{show_inf_plot}{print inference plot, set to verbose by default}

\item{show_res}{print results, set to verbose by default}
}
\description{
Helper for the \code{inference()} function
}
